+++ Dockerfile.api
@@ -6,5 +6,4 @@ WORKDIR /api
 ENV FLASK_ENV production
 
 EXPOSE 5000 1883
-#CMD [ "python3", "-m", "flask", "run" ]
 CMD ["gunicorn", "-b", ":5000", "app:app"]
