+++ Dockerfile.client
@@ -1,4 +1,3 @@
-# Build step #1: build the React front end
 FROM node:16-alpine as build-step
 ENV PATH /frontend/node_modules/.bin:$PATH
