+++ frontend/package.json
@@ -43,5 +43,5 @@
     "postcss": "^8.4.13",
     "tailwindcss": "^3.0.24"
   },
-  "proxy": "http://app:5000/"
+  "proxy": "http://localhost:5000/"
 }
