+++ frontend/src/App.js
@@ -1,12 +1,12 @@
 import React from 'react';
-import TilesCust from './Components/TilesCust'
+import Homepage from './Components/Homepage'
 
 function App() {
 
     return (
         <div>
             <h1 className="text-white text-3xl text-center mt-4 mb-4">Smart Home</h1>
-            <TilesCust/>
+            <Homepage/>
         </div>
     );
 }
